- Modifications to GTK::Application for GTK::Builder support. - Changes to GTK::Builder and GTK::Raw::Builder. - Added new passthrough constructor, but now constructing from GtkWidet and GtkWindow can be distinctive - Changes to t/40-builder.t
Comments: GTK::Builder support was attempted too early. There are still
differences that need to be taken into account. For one thing, ANYTHING
returned from GtkBuilder will be a GOBJECT, and NOT a GtkWidget, no
matter what is done in GTK::Raw::Builder.

Secondly, the pointer-to-object resolution mechanism should be put into
place.

The basic design SEEMS to be sound, but until we can get GTK to
recognize the objects pulled from GtkBuilder, we cannot verify proper
operation.

GtkBuilder may need to go back to the LAST thing that is supported. In
the meantime, the original work can be continued.