- Finishes the aliasing sprint, for now. All current modules have proper aliasing. In many cases, the decision was made to rename methods to prevent methods conflicting with signal names. After much thought, it was thought that the methods can be renamed back to the GTK compliant names with a different signal handling method. Therefore, in a future revision, it is proposed that all signals are accessed via a "signal" method that returns a hash.
Currently, to tap a signal, one does the following:

$button.clicked.tap({...})

This will then become

$button.signal<clicked>.tap({...})

This will then free up the method "clicked" for future use. As of right
now, if there is a function to emit a signal on an object, we use:

$button.emit-clicked

Such a redesign will allow for

$button.clicked

to emit the clicked event.

Keep this in mind for the future.