- Updated BuildList, META6.json and GTK::Builder::MRO - build.sh now runs scripts/dependencies.pl6 prior to building - Circumstances have now forced us to require cheating, therefore a prototype CStruct for GObject has been created. This is ONLY for use in Widget-based objects, since they now need it for proper functionality. - Fixes error with GTK::Raw::Data routines. The end result of this is that GTK::Widget typing now works. - Modification to warning in GTK::Roles::Data when using .setType - GTK::Toolbar wasn't properly implementing the new roles assigned to it. This has now been fixed. - Improved destruction logic in GTK::Widget - Now using GTK::Widget.CreateObject to instantiate signal handler objects where the original object was NOT available via closure. in t/35-dnd.t
+ NOTE!! -- Object creation via identity constructor in ALL widgets is
faulty. This includes objects created via GTK::Widget.CreateObject. The
problem stems from the fact that while the objects are created properly,
the reference count on the object pointer is not incremented. This is a
problem since GTK::Widget.DESTROY will DECREMENT the reference count
when the OBJECT goes out of scope. This means that ALL identity
constructors will now need to perform a .upref upon creation. This is a
PRIORITY refactoring and will occur over the next few commits.