- Add NativeCall sub for no-arg, integer return version of g_idle_add (g_idle_add_rint).
+ This kicks over the fact that GCallBacks are meant to be mutable 
handlers. Currently the only way to handle that is related to how we do 
signals: multiple NC subs to one symbol. While that worked for symbols, 
that has huge limitations to CallBacks, and we won't know what to define 
ahead of time. It would be one thing to do this for EVERY SIGNAL HANDLER 
we have, but what about future callbacks?